block: replace `Base` with `Header`
