+++ pkg/network/payload/headers_test.go
@@ -35,12 +35,10 @@ func newTestHeaders(n int) *Headers {
 
 	for i := range headers.Hdrs {
 		headers.Hdrs[i] = &block.Header{
-			Base: block.Base{
-				Index: uint32(i + 1),
-				Script: transaction.Witness{
-					InvocationScript:   []byte{0x0},
-					VerificationScript: []byte{0x1},
-				},
+			Index: uint32(i + 1),
+			Script: transaction.Witness{
+				InvocationScript:   []byte{0x0},
+				VerificationScript: []byte{0x1},
 			},
 		}
 	}
